Readability refactoring
